initial move from Primer::HiddenTextExpander to Primer::Alpha::HiddenTextExpander
