+++ .changeset/dry-cobras-build.md
@@ -2,4 +2,4 @@
 '@primer/view-components': patch
 ---
 
-moving Primer::HiddenTextExpander to Primer::Alpha::HiddenTextExpander and replacing the original with a deprecated version
+moving Primer::Alpha::HiddenTextExpander to Primer::Alpha::HiddenTextExpander and replacing the original with a deprecated version
