+++ lib/primer/view_components/linters/helpers/deprecated_components_helpers.rb
@@ -7,6 +7,7 @@ module ERBLint
       module DeprecatedComponentsHelpers
         # If there is no alternative to suggest, set the value to nil
         COMPONENT_TO_USE_INSTEAD = {
+          "Primer::HiddenTextExpander" => "Primer::Alpha::HiddenTextExpander",
           "Primer::HeadingComponent" => "Primer::Beta::Heading",
           "Primer::ButtonGroup" => "Primer::Beta::ButtonGroup",
           "Primer::CloseButton" => "Primer::Beta::CloseButton",
