+++ static/arguments.yml
@@ -22,6 +22,21 @@
     type: Hash
     default: N/A
     description: "[System arguments](/system-arguments)"
+- component: HiddenTextExpander
+  source: https://github.com/primer/view_components/tree/main/app/components/primer/alpha/hidden_text_expander.rb
+  parameters:
+  - name: inline
+    type: Boolean
+    default: "`false`"
+    description: Whether or not the expander is inline.
+  - name: button_arguments
+    type: Hash
+    default: "`{}`"
+    description: "[System arguments](/system-arguments) for the button element."
+  - name: system_arguments
+    type: Hash
+    default: N/A
+    description: "[System arguments](/system-arguments)"
 - component: Layout
   source: https://github.com/primer/view_components/tree/main/app/components/primer/alpha/layout.rb
   parameters:
@@ -745,21 +760,6 @@
     type: Boolean
     default: "`false`"
     description: Whether or not the button is inline.
-  - name: system_arguments
-    type: Hash
-    default: N/A
-    description: "[System arguments](/system-arguments)"
-- component: HiddenTextExpander
-  source: https://github.com/primer/view_components/tree/main/app/components/primer/hidden_text_expander.rb
-  parameters:
-  - name: inline
-    type: Boolean
-    default: "`false`"
-    description: Whether or not the expander is inline.
-  - name: button_arguments
-    type: Hash
-    default: "`{}`"
-    description: "[System arguments](/system-arguments) for the button element."
   - name: system_arguments
     type: Hash
     default: N/A
