+++ static/audited_at.json
@@ -2,6 +2,7 @@
   "Primer::Alpha::AutoComplete": "",
   "Primer::Alpha::AutoComplete::Item": "",
   "Primer::Alpha::ButtonMarketing": "",
+  "Primer::Alpha::HiddenTextExpander": "",
   "Primer::Alpha::Layout": "",
   "Primer::Alpha::Layout::Main": "",
   "Primer::Alpha::Layout::Sidebar": "",
