+++ static/statuses.json
@@ -2,6 +2,7 @@
   "Primer::Alpha::AutoComplete": "deprecated",
   "Primer::Alpha::AutoComplete::Item": "deprecated",
   "Primer::Alpha::ButtonMarketing": "alpha",
+  "Primer::Alpha::HiddenTextExpander": "alpha",
   "Primer::Alpha::Layout": "alpha",
   "Primer::Alpha::Layout::Main": "alpha",
   "Primer::Alpha::Layout::Sidebar": "alpha",
@@ -51,7 +52,7 @@
   "Primer::FlexItemComponent": "deprecated",
   "Primer::HeadingComponent": "deprecated",
   "Primer::HellipButton": "alpha",
-  "Primer::HiddenTextExpander": "alpha",
+  "Primer::HiddenTextExpander": "deprecated",
   "Primer::IconButton": "beta",
   "Primer::Image": "alpha",
   "Primer::ImageCrop": "alpha",
