+++ stories/primer/alpha/hidden_text_expander_stories.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-class Primer::HiddenTextExpanderStories < ViewComponent::Storybook::Stories
+require "primer/alpha/hidden_text_expander"
+
+class Primer::Alpha::HiddenTextExpanderStories < ViewComponent::Storybook::Stories
   layout "storybook_preview"
 
   story(:hidden_text_expander) do
