+++ test/components/alpha/hidden_text_expander_test.rb
@@ -2,11 +2,11 @@
 
 require "test_helper"
 
-class PrimerHiddenTextExpanderTest < Minitest::Test
+class PrimerAlphaHiddenTextExpanderTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_renders
-    render_inline(Primer::HiddenTextExpander.new("aria-label": "No effect"))
+    render_inline(Primer::Alpha::HiddenTextExpander.new("aria-label": "No effect"))
 
     assert_selector("span.hidden-text-expander") do
       assert_selector("button[type='button'][aria-expanded='false'].ellipsis-expander", text: "…")
@@ -14,7 +14,7 @@ class PrimerHiddenTextExpanderTest < Minitest::Test
   end
 
   def test_renders_inline
-    render_inline(Primer::HiddenTextExpander.new(inline: true, "aria-label": "No effect"))
+    render_inline(Primer::Alpha::HiddenTextExpander.new(inline: true, "aria-label": "No effect"))
 
     assert_selector("span.hidden-text-expander.inline") do
       assert_selector("button[type='button'][aria-expanded='false'].ellipsis-expander", text: "…")
@@ -22,7 +22,7 @@ class PrimerHiddenTextExpanderTest < Minitest::Test
   end
 
   def test_renders_button_custom_classes
-    render_inline(Primer::HiddenTextExpander.new("aria-label": "No effect", button_arguments: { classes: "custom-class" }))
+    render_inline(Primer::Alpha::HiddenTextExpander.new("aria-label": "No effect", button_arguments: { classes: "custom-class" }))
 
     assert_selector("span.hidden-text-expander") do
       assert_selector("button[type='button'][aria-expanded='false'].ellipsis-expander.custom-class", text: "…")
@@ -30,27 +30,27 @@ class PrimerHiddenTextExpanderTest < Minitest::Test
   end
 
   def test_renders_aria_label
-    render_inline(Primer::HiddenTextExpander.new("aria-label": "Custom aria label"))
+    render_inline(Primer::Alpha::HiddenTextExpander.new("aria-label": "Custom aria label"))
 
     assert_selector("button[aria-label='Custom aria label']")
   end
 
   def test_renders_aria_label_provided_as_object
-    render_inline(Primer::HiddenTextExpander.new(aria: { label: "Custom aria label" }))
+    render_inline(Primer::Alpha::HiddenTextExpander.new(aria: { label: "Custom aria label" }))
 
     assert_selector("button[aria-label='Custom aria label']")
   end
 
   def test_raises_if_no_aria_label_is_provided
     err = assert_raises ArgumentError do
-      render_inline(Primer::HiddenTextExpander.new)
+      render_inline(Primer::Alpha::HiddenTextExpander.new)
     end
 
     assert_equal("`aria-label` is required.", err.message)
   end
 
   def test_does_not_render_content
-    render_inline(Primer::HiddenTextExpander.new("aria-label": "No effect")) { "content" }
+    render_inline(Primer::Alpha::HiddenTextExpander.new("aria-label": "No effect")) { "content" }
 
     refute_text("content")
   end
